cont. test refactoring; resolved conflicts
