+++ api/v1/models/User.js
@@ -81,7 +81,6 @@ User.prototype.setPassword = function (password) {
 		.hashAsync(password, SALT_ROUNDS)
 		.bind(this)
 		.then(function (p) {
-			console.log(p);
 			return _Promise.resolve(this.set({ password: p }));
 		});
 };
